Moved init from node api to commands
